Complete
This is now a complete game that  will take aplhapets using buttons in windows form from the user check weather that letter is part of the randomly generated word if it is then will display the alphabet in its correct position if not then indicate that guess is wrong by changing color of button. If all guesses are correct display you win if you run out of chances display you lose.